backend/requirements.txt
# List of backend dependencies
# Flask API and extensions
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# Data processing
pandas==2.0.3
numpy==1.24.4
# NLP & ML
nltk==3.8.1
textblob==0.17.1
scikit-learn==1.3.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Database
sqlite3client==0.0.1
# Utils
python-dateutil==2.8.2
pytz==2023.3